replace return type int to long

+++ 0x02-functions_nested_loops/103-fibonacci.c
@@ -8,7 +8,7 @@
 
 int main(void)
 {
-int even_sum = 0;
+long even_sum = 0;
 long a = 1;
 long b = 1;
 long sum;
